AI 功能上线

+++ internal/ai/internal/web/vo.go
@@ -22,8 +22,8 @@ type JDResponse struct {
 }
 
 type JDEvaluation struct {
-	Score    int    `json:"score"`
-	Analysis string `json:"analysis"`
+	Score    float64 `json:"score"`
+	Analysis string  `json:"analysis"`
 }
 
 type Config struct {
